drive_bot_final/requirements.txt

aiogram==3.4.1
python-docx>=1.1.0
pdfplumber>=0.10
python-Levenshtein>=0.22
transliterate>=1.10
google-api-python-client>=2.126
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.0
markdownify>=0.11
celery[redis]>=5.3
orjson>=3.10
python-json-logger>=2
spacy>=3.7
# ru-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_lg-3.7.0/ru_core_news_lg-3.7.0-py3-none-any.whl
pymupdf>=1.23
pdf2image>=1.17
pytesseract>=0.3
pillow>=10
redis[asyncio]>=4.6
